move assets/ out of store/
