+++ core/assets/currencies.go
