+++ core/store/models/job_run.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"time"
 
+	"chainlink/core/assets"
 	clnull "chainlink/core/null"
-	"chainlink/core/store/assets"
 	"chainlink/core/utils"
 
 	"github.com/ethereum/go-ethereum/common"
