+++ core/store/models/log_events.go
@@ -6,9 +6,9 @@ import (
 	"fmt"
 	"math/big"
 
+	"chainlink/core/assets"
 	"chainlink/core/eth"
 	"chainlink/core/logger"
-	"chainlink/core/store/assets"
 	"chainlink/core/utils"
 
 	ethereum "github.com/ethereum/go-ethereum"
