+++ core/web/config_controller_test.go
@@ -6,8 +6,8 @@ import (
 	"testing"
 	"time"
 
+	"chainlink/core/assets"
 	"chainlink/core/internal/cltest"
-	"chainlink/core/store/assets"
 	"chainlink/core/store/orm"
 	"chainlink/core/store/presenters"
