+++ core/web/transfer_controller_test.go
@@ -6,9 +6,9 @@ import (
 	"net/http"
 	"testing"
 
+	"chainlink/core/assets"
 	"chainlink/core/eth"
 	"chainlink/core/internal/cltest"
-	"chainlink/core/store/assets"
 	"chainlink/core/store/models"
 
 	"github.com/ethereum/go-ethereum/common"
